input: sec_ts: Update input open/close control logic
Switch to a mixed mode, since the logic that swapping the input status
is not that guaranteed. The new logic combines both fb and node status
status to update the open/close.
 - Only perform open if last fb status and current status are both
   close. This is to workaround the case when ambient display is
   triggered in screen off state.
 - In all other cases we still need to perform a open/close to update
   the fingerprint status.
This fixes the input when ambient display is triggered in screen off
state. However, during AOD it's still broken and is currently not
fixable, due to no change to both fb and node status.